DeleteBookConfirm component separated
