+++ src/components/MyBooks/DeleteBookConfirmation/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DeleteBookConfirmation";
