+++ src/hooks/useDeleteBookQuery.tsx
@@ -1,6 +1,5 @@
 import api from "api";
 import { useMutation } from "react-query";
-import { useState } from "react";
 type props = {
 	bookId: number;
 	onSuccess?: Function;
