+++ src/pages/BookDisplay/BookDisplay.tsx
@@ -57,7 +57,6 @@ const BookDisplay: React.FC = () => {
 					</h1>
 				);
 			case "error":
-				// TODO: make a Route named /notfound for SbG
 				history.push("/notfound");
 				break;
 
@@ -164,7 +163,6 @@ const BookDisplay: React.FC = () => {
 								<img
 									alt="ecommerce"
 									className="hidden object-cover object-center mx-auto mt-4 rounded md:inline-flex md:w-1/5 md:h-1/4"
-									// NEXTCOMMIT: src={data?.data.data.book.images[0].url}
 									src={data?.data.data.book.image.url}
 								/>
 							</div>
